refactored some region code, there might still be artifacts
